Revert "TS migration"
