--- src/types/prompts.ts
@@ -1,3 +1,0 @@
-export type TPromptFunctions = (props: any) => string;
-
-export type TPromptTemplates = Record<string, TPromptFunctions>;
