--- src/types/store.ts
@@ -1,5 +1,0 @@
-/**
- * ### Store types
- * @todo Implement various stores later on.
- */
-export type TStore = any;
