+++ src/utils/telemetry.js
